requirements.txt
#torch=1.7.1
torchtext==0.8.1
torchsummary==1.5.1
tensorboard==2.4.1
matplotlib==3.3.4
transformers==4.3.2
pandas==1.1.5
sentencepiece==0.1.95
datasets==1.2.1
guildai==0.7.2
sklearn
